Rename files from CamelCase to snake_case and update imports
- Renamed all Python files to follow snake_case naming convention
- Updated imports to use relative imports with the format from .module_name import ClassName
- Updated doctests to use new import format
- Created __init__.py with direct exports to maintain API compatibility
