+++ conjure/input_data_py_read.py
@@ -1,8 +1,8 @@
 import pandas as pd
 from typing import List
 
-from edsl.conjure.InputData import InputDataABC
-from edsl.conjure.utilities import convert_value
+from .input_data import InputDataABC
+from .utilities import convert_value
 from edsl.utilities.utilities import is_valid_variable_name
 
 try:
