+++ conjure/raw_question.py
@@ -3,7 +3,7 @@ from typing import List, Optional, Union
 from edsl.questions import QuestionBase
 from edsl import Question
 
-from conjure.utilities import convert_value
+from .utilities import convert_value
 
 
 @dataclass
